+++ CHANGELOG.md
@@ -7,6 +7,10 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 ## [Unreleased]
 ### Added
  * Added a compiler metadata flag for suppressing warnings when Var indirection is unavoidable (#1052)
+ * Added the `--emit-generated-python` CLI argument to control whether generated Python code strings are stored by the runtime for each compiled namespace (#1045)
+
+### Changed
+ * The compiler will issue a warning when adding any alias that might conflict with any other alias (#1045)
 
 ### Fixed
  * Basilisp now respects the value of Python's `sys.dont_write_bytecode` flag when generating bytecode (#1054)